Additional methods structured
