+++ Lusas_Adapter/Lusas_Adapter.csproj
@@ -90,7 +90,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CRUD\Create\Additional_Methods.cs" />
     <Compile Include="CRUD\Create\BarPointLoad.cs" />
     <Compile Include="CRUD\Create\BarDistributedLoad.cs" />
     <Compile Include="CRUD\Create\SurfaceMesh.cs" />
@@ -113,7 +112,13 @@
     <Compile Include="CRUD\Create\TemperatureLoad.cs" />
     <Compile Include="CRUD\Create\_Create.cs" />
     <Compile Include="CRUD\Delete\Delete.cs" />
-    <Compile Include="CRUD\NextId.cs" />
+    <Compile Include="Adapter\NextId.cs" />
+    <Compile Include="CRUD\Private Helpers\AssignObjectSet.cs" />
+    <Compile Include="CRUD\Private Helpers\CreateTags.cs" />
+    <Compile Include="CRUD\Private Helpers\GetAssignedObjects.cs" />
+    <Compile Include="CRUD\Private Helpers\GetDistinct.cs" />
+    <Compile Include="CRUD\Private Helpers\GetLoadAssignments.cs" />
+    <Compile Include="CRUD\Private Helpers\NameSearch.cs" />
     <Compile Include="CRUD\Read\Properties\1DMeshSettings.cs" />
     <Compile Include="CRUD\Read\Properties\2DMeshSettings.cs" />
     <Compile Include="CRUD\Read\Properties\2DProperties.cs" />
@@ -144,7 +149,7 @@
     <Compile Include="CRUD\Read\Properties\Tags.cs" />
     <Compile Include="CRUD\Read\_Read.cs" />
     <Compile Include="CRUD\Update\UpdateObjects.cs" />
-    <Compile Include="LusasAdapter.cs" />
+    <Compile Include="Adapter\LusasAdapter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Types\Comparer.cs" />
     <Compile Include="Types\DependencyTypes.cs" />
