+++ Lusas_oM/Settings/LibrarySettings.cs
