+++ Lusas_oM/Structural/Properties/MeshSettings1D.cs
@@ -1,0 +1,16 @@
+﻿using BH.oM.Base;
+
+namespace BH.oM.Adapters.Lusas
+{
+    public class MeshSettings1D : BHoMObject
+    {
+        /***************************************************/
+        /****            Public Properties              ****/
+        /***************************************************/
+
+        public Split1D SplitMethod { get; set; }
+        public double SplitParameter { get; set; }
+
+        /***************************************************/
+    }
+}
