+++ Lusas_oM/Structural/Properties/MeshSettings2D.cs
@@ -2,18 +2,6 @@
 
 namespace BH.oM.Adapters.Lusas
 {
-    public class MeshSettings1D : BHoMObject
-    {
-        /***************************************************/
-        /****            Public Properties              ****/
-        /***************************************************/
-
-        public Split1D SplitMethod { get; set; }
-        public double SplitParameter { get; set; }
-
-        /***************************************************/
-    }
-
     public class MeshSettings2D : BHoMObject
     {
         /***************************************************/
